Moved NegZInt under NegInts.
