+++ dotty/scalactic/src/main/scala/org/scalactic/opaquetypes/NegInts.scala
@@ -17,13 +17,20 @@ package org.scalactic.opaquetypes
 
 import org.scalactic.Resources
 
-opaque type NegZInt = Int
+object NegInts {
+  opaque type NegInt = Int
+  object NegInt {
+    def ensuringValid(i: Int): NegInt = 
+      if (i >= 0) 
+        throw new AssertionError(Resources.invalidNegInt)
+      else i
+  }
 
-object NegZInt {
-
-  def ensuringValid(i: Int): NegZInt = 
-    if (i > 0) 
-      throw new AssertionError(Resources.invalidNegZInt)
-    else i
-
+  opaque type NegZInt = Int
+  object NegZInt {
+    def ensuringValid(i: Int): NegZInt = 
+      if (i > 0) 
+        throw new AssertionError(Resources.invalidNegZInt)
+      else i
+  }
 }
